+++ www/go/core/views/extjs3/Entities.js
@@ -28,7 +28,17 @@ go.Entities = (function () {
 				}
 				
 				if(entity.customFields) {
-					entity.filters = entity.filters.concat(go.customfields.CustomFields.getFilters(entity.name));
+					var existingNames = entity.filters.column("name");
+					var customFieldFilters = go.customfields.CustomFields.getFilters(entity.name);
+					customFieldFilters = customFieldFilters.filter(function(f) {
+						var exists = existingNames.indexOf(f.name) > -1;
+						if(exists) {
+							console.warn("Custom field name " + f.name+ " can't be filtered as the name conflicts with an existing filter for entity " + entity.name);
+						}
+						return !exists;
+
+					});
+					entity.filters = entity.filters.concat(customFieldFilters);
 				}
 
 				entity.filters =  go.util.Filters.normalize(entity.filters);