+++ www/modules/calendar/model/Event.php
@@ -56,6 +56,7 @@ use DateTimeZone;
 use GO\Calendar\Model\Exception;
 use GO;
 use GO\Base\Util\StringHelper;
+use go\core\model\Module;
 use Sabre;
 use Swift_Attachment;
 use Swift_Mime_ContentEncoder_PlainContentEncoder;
@@ -330,7 +331,7 @@ class Event extends \GO\Base\Db\ActiveRecord {
 
 	
 		if(!$exception->save()){
-			throw new Exception("Event exception not saved: ".var_export($exception->getValidationErrors(), true));
+			throw new \Exception("Event exception not saved: ".var_export($exception->getValidationErrors(), true));
 		}
 			
 		
@@ -969,7 +970,7 @@ class Event extends \GO\Base\Db\ActiveRecord {
 
 					$message = \GO\Base\Mail\Message::newInstance(
 										$subject
-										)->setFrom(\GO::user()->email, \GO::user()->name)
+										)->setFrom(go()->getSettings()->systemEmail, go()->getSettings()->title)
 										->addTo($adminUser->email, $adminUser->name);
 
 					$message->setHtmlAlternateBody($body);					
@@ -1014,7 +1015,7 @@ class Event extends \GO\Base\Db\ActiveRecord {
 
 				$message = \GO\Base\Mail\Message::newInstance(
 									$subject
-									)->setFrom(\GO::user()->email, \GO::user()->name)
+									)->setFrom(go()->getSettings()->systemEmail, go()->getSettings()->title)
 									->addTo($this->user->email, $this->user->name);
 
 				$message->setHtmlAlternateBody($body);					
@@ -2034,7 +2035,7 @@ The following is the error message:
 										);
 						$message = \GO\Base\Mail\Message::newInstance(
 														$mailSubject
-														)->setFrom(\GO::config()->webmaster_email, \GO::config()->title)
+														)->setFrom(go()->getSettings()->systemEmail, go()->getSettings()->title)
 														->addTo($this->calendar->user->email);
 
 						$message->setHtmlAlternateBody(nl2br($body));
@@ -2443,9 +2444,16 @@ The following is the error message:
 	 * 
 	 * @return Participant
 	 */
-	public function getParticipantOfCalendar(){
+	public function getParticipantOfCalendar() {
+
+		$aliases = GO\Email\Model\Alias::model()->find(
+				GO\Base\Db\FindParams::newInstance()
+					->select('email')
+					->permissionLevel(GO\Base\Model\Acl::WRITE_PERMISSION, $this->calendar->user_id)
+				)->fetchAll(\PDO::FETCH_COLUMN, 0);
+
 		return Participant::model()->findSingleByAttributes(array(
-				'user_id'=>$this->calendar->user_id,
+				'email' => $aliases,
 				'event_id'=>$this->id
 		));
 	}
@@ -2546,7 +2554,7 @@ The following is the error message:
 
 		$message->setHtmlAlternateBody($body);
 
-		\GO\Base\Mail\Mailer::newGoInstance()->send($message);
+		$this->getUserMailer()->send($message);
 
 	}
 	
@@ -2613,8 +2621,8 @@ The following is the error message:
 			// Set back the original language
 			if($language !== false)
 				\GO::language()->setLanguage($language);
-			
-			\GO\Base\Mail\Mailer::newGoInstance()->send($message);
+
+			$this->getUserMailer()->send($message);
 		}
 
 		return true;
@@ -2730,10 +2738,9 @@ The following is the error message:
 					if($language !== false)
 						\GO::language()->setLanguage($language);
 
-					if(!\GO\Base\Mail\Mailer::newGoInstance()->send($message)) {
+					if(!$this->getUserMailer()->send($message)) {
 						throw new \Exception("Failed to send invite");
 					}
-
 					
 				}
 				
@@ -2742,6 +2749,23 @@ The following is the error message:
 			unset(\GO::session()->values['new_participant_ids']);
 			
 			return true;
+	}
+
+	/**
+	 *
+	 * @return GO\Base\Mail\Mailer
+	 */
+	private function getUserMailer() {
+
+		if(Module::isInstalled('legacy', 'email')) {
+			$account = GO\Email\Model\Account::model()->findByEmail($this->user->email)->findSingle();
+			if($account) {
+				$transport = GO\Email\Transport::newGoInstance($account);
+				return \GO\Base\Mail\Mailer::newGoInstance($transport);
+			}
+		}
+
+		return \GO\Base\Mail\Mailer::newGoInstance();
 	}
 	
 	public function resourceGetEventCalendarName() {