+++ www/modules/chat/language/de.php
@@ -1,4 +1,0 @@
-<?php
-return array (
-  'description' => 'Verbindet mit einem XMPP Server und authentifiziert die {product_name} Benutzer.',
-);