+++ www/modules/chat/language/es.php
@@ -1,5 +1,0 @@
-<?php
-return array (
-  'name' => 'Módulo de Chat',
-  'description' => 'Permite la conexión a un servidor XMPP y compartir los mismos usuarios que {product_name}. Por ejemplo Prosody con autenticación desde {product_name} o desde un backoffice LDAP.',
-);